requirements.txt
tensorflow==1.7.0
tqdm
Flask
tensorflow-hub==0.1.0
numpy
google-images-download
selenium
python-magic
